refactor: decouple resolvers from module_loader.rs for standalone use
